+++ examples/src/Examples/Fragments/InlineImports.hs
@@ -47,7 +47,7 @@ struct Params {
   color: vec4<f32>;
 };
 
-@group(3) @binding(0)
+@group(0) @binding(0)
 var<uniform> params: Params;
 
 @fragment